requirements.txt
discord.py
numpy
PyNaCl==1.3.0